requirements3.txt
#
# This file is autogenerated by pip-compile
#
blinker==1.4
boto3==1.4.5
botocore==1.5.95          # via boto3, s3transfer
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
coverage==4.5.4           # via pytest-cov
docutils==0.15.2          # via botocore
flask-script==2.0.5
flask==1.1.1
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.9.0
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
psutil==5.6.3             # via rainbow-saddle
py==1.8.0                 # via pytest
pytest-cov==2.5.1
pytest-mock==1.1
pytest==3.1.1
python-dateutil==2.8.0    # via botocore
python-json-logger==0.1.11
python-redis-lock==3.2.0
pyyaml==3.12
rainbow-saddle==0.4.0
redis==2.10.6
requests==2.22.0          # via slackclient
s3transfer==0.1.13        # via boto3
six==1.12.0               # via python-dateutil, slackclient, websocket-client
slackclient==1.0.7
statsd==3.2.1
urllib3==1.25.3           # via requests
websocket-client==0.56.0  # via slackclient
werkzeug==0.15.5          # via flask

pip==9.0.3
setuptools==41.1.0
